+++ src/ParameterAccess.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bakame\Http\StructuredFields;
+
+/**
+ * @property-read MemberOrderedMap<string, StructuredField> $parameters
+ */
+interface ParameterAccess
+{
+}