+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 ### Added
 
 - Introduce the`MemberContainer` interface.
+- Introduce the`ParameterAccess` interface.
 - `OrderedList`, `InnerList` now implements the `MemberList` interface.
 - `Parameters` and `Dictionnary` now implements the PHP `MemberOrderedMap` interface.
 - `Token::value` is a readonly property.